Добавляет модуляцию
